requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
envparse==0.2.0
exceptiongroup==1.1.0
fastapi==0.89.1
filelock==3.9.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.17
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-dotenv==0.5.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.19.0
